// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        time_entry: {
            id: number;
            project?: components["schemas"]["id_name"];
            issue?: {
                id: number;
            };
            user?: components["schemas"]["id_name"];
            activity?: components["schemas"]["id_name"];
            hours: number;
            comments: string | null;
            /** Format: date */
            spent_on: string;
            /** Format: date-time */
            created_on: string;
            /** Format: date-time */
            updated_on: string;
            custom_fields?: components["schemas"]["custom_field_value"][];
        };
// ...
    };
// ...
}